adjust did-document trait
